# backend/openapi.yaml
openapi: 3.0.3
info:
  title: Yercaud Business Directory API
  version: 0.1.0
  description: >
    Backend contract for the public frontend and Admin panel. Auth is
    httpOnly session cookies (ADR: session cookies over JWT). See
    /CONTEXT.md and /docs/adr/ for the domain model and design decisions
    this contract implements.
servers:
  - url: http://localhost:4000/v1
    description: Local dev
security:
  - cookieAuth: []
...
paths:
...
  /blog-categories:
    get:
      tags: [Blog]
      security: []
      summary: List blog categories
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/NamedLookup' }
    post:
      tags: [Blog]
      summary: Create a blog category (Super Admin) — slug generated from the name
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/NamedLookupInput' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/NamedLookup' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }

  # Phase 8: the original contract had list+create only for blog categories,
  # FAQ categories, and FAQs. These per-resource paths add the rest.
  /blog-categories/{categoryId}:
    parameters: [{ $ref: '#/components/parameters/CategoryId' }]
    patch:
      tags: [Blog]
      summary: Rename a blog category (Super Admin) — re-slugs from the new name
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/NamedLookupInput' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/NamedLookup' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
    delete:
      tags: [Blog]
      summary: Delete a blog category (Super Admin) — 409 if it still has posts (ADR 0009 RESTRICT)
      responses:
        '204': { description: Deleted }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
...
  /blog-posts/{postId}/cover-image:
    parameters: [{ $ref: '#/components/parameters/PostId' }]
    post:
      tags: [Blog]
      summary: Upload a blog post's cover image (stored on local disk, same as Listing images)
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file: { type: string, format: binary }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/BlogPost' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /blog-posts/{postId}/comments:
    parameters: [{ $ref: '#/components/parameters/PostId' }]
    get:
      tags: [Blog]
      security: []
      summary: List a post's approved comments
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/BlogComment' }
    post:
      tags: [Blog]
      summary: Post a comment — starts 'pending' (same moderation shape as Review)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [body]
              properties: { body: { type: string } }
      responses:
        '201': { description: "Created, pending approval" }

  # Phase 8: the cross-post moderation queue, mirroring GET /reviews.
  /blog-comments:
    get:
      tags: [Blog]
      summary: List comments across all posts and statuses (Super Admin moderation queue)
      parameters:
        - name: status
          in: query
          schema: { type: string, enum: [pending, approved, rejected] }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/BlogComment' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /blog-comments/{commentId}/approve:
    parameters: [{ $ref: '#/components/parameters/CommentId' }]
    post:
      tags: [Blog]
      summary: Approve a pending comment (Super Admin)
      responses:
        '200':
          description: Approved
          content:
            application/json:
              schema: { $ref: '#/components/schemas/BlogComment' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }

  # Phase 8: blog_comments.status has a 'rejected' value the contract gave no
  # way to reach.
  /blog-comments/{commentId}/reject:
    parameters: [{ $ref: '#/components/parameters/CommentId' }]
    post:
      tags: [Blog]
      summary: Reject a pending comment (Super Admin)
      responses:
        '200':
          description: Rejected
          content:
            application/json:
              schema: { $ref: '#/components/schemas/BlogComment' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: session_id
  parameters:
...
    CategoryId:
      name: categoryId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
    CommentId:
      name: commentId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
    PostId:
      name: postId
      in: path
      required: true
      schema: { type: string, format: uuid }
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
    Forbidden:
      description: Not permitted (RBAC)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    NotFound:
      description: Not found
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Conflict:
      description: Conflicts with an existing resource or constraint
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
  schemas:
    Error:
      type: object
      properties:
        error: { type: string }
        message: { type: string }
...
    NamedLookup:
      type: object
      required: [id, name]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        # Present on Categories and Locations; absent on the plain lookups
        # (amenities, attribute tags, property types, languages), which have no
        # slug column.
        slug: { type: string }
    NamedLookupInput:
      type: object
      required: [name]
      properties:
        name: { type: string }
...
    BlogPostSummary:
      type: object
      # Declared required because the response always carries them — same
      # reasoning as ListingSummary above (Phase 10): without this every
      # generated field is `| undefined`, which defeats generated types.
      required:
        - id
        - title
        - slug
        - excerpt
        - coverImage
        - categoryId
        - authorId
        - authorName
        - authorAvatarUrl
        - authorBio
        - readingTimeMinutes
        - status
        - publishedAt
        - createdAt
        - viewCount
        - tags
      properties:
        id: { type: string, format: uuid }
        title: { type: string }
        slug: { type: string }
        excerpt: { type: string, nullable: true }
        coverImage: { type: string, nullable: true }
        categoryId: { type: string, format: uuid }
        authorId: { type: string, format: uuid }
        # author_id is NOT NULL, so these are always populated — unlike
        # Review's authorName, which is optional because that module has a
        # second query variant that doesn't join users at all.
        authorName: { type: string }
        authorAvatarUrl: { type: string, nullable: true }
        authorBio: { type: string, nullable: true }
        readingTimeMinutes: { type: integer, nullable: true }
        status: { type: string, enum: [draft, pending, published] }
        publishedAt: { type: string, format: date-time, nullable: true }
        createdAt: { type: string, format: date-time }
        viewCount: { type: integer }
        tags: { type: array, items: { type: string } }
    BlogPost:
      allOf:
        - $ref: '#/components/schemas/BlogPostSummary'
        - type: object
          # Declared required because the response always carries them — same
          # reasoning as BlogPostSummary/BlogComment above (toBlogPost always
          # sets every one of these, nullable fields included, to null rather
          # than omitting them).
          required:
            - body
            - seoTitle
            - seoDescription
            - ogImage
            - twitterCard
            - relatedPostIds
            - placeListingIds
          properties:
            body: { type: string }
            seoTitle: { type: string, nullable: true }
            seoDescription: { type: string, nullable: true }
            # FR189, closed in Phase 10. Phase 9 exposed these on Listings, where
            # the columns already existed unused, and left Blog Posts out because
            # they needed a migration — noting the asymmetry it created. A Blog
            # Post is the thing most likely to be shared into WhatsApp, so the
            # asymmetry was backwards.
            ogImage: { type: string, nullable: true }
            twitterCard:
              type: string
              nullable: true
              enum: [summary, summary_large_image, app, player]
            relatedPostIds: { type: array, items: { type: string, format: uuid } }
            placeListingIds: { type: array, items: { type: string, format: uuid } }
    # `slug`, `readingTimeMinutes`, `authorId`, `status`, and `publishedAt` are
    # all absent by design: the first two are derived server-side from title
    # and body, authorId comes from the session, and status only moves via
    # PATCH or /publish. A new post always starts as 'draft'.
...
    BlogComment:
      type: object
      # Declared required because the response always carries them — same
      # reasoning as ListingSummary/BlogPostSummary (a pre-existing gap here
      # left every field `| undefined`, found while wiring issue #24).
      required: [id, postId, userId, body, likeCount, status, createdAt]
      properties:
        id: { type: string, format: uuid }
        postId: { type: string, format: uuid }
        userId: { type: string, format: uuid }
        body: { type: string }
        # Read-only: nothing in this API likes a comment, so this is always 0.
        likeCount: { type: integer }
        status: { type: string, enum: [pending, approved, rejected] }
        createdAt: { type: string, format: date-time }

    # Phase 8: NOT a NamedLookup, which the contract previously claimed —
    # faq_categories has a sort_order and no slug, the opposite of that shape.
...